Basic styling for RC1 Workouts
